adding tokenized validation to settings page to defeat csrf attack
git-svn-id: http://svn.concrete5.org/svn/concrete5@316 b0551a0c-1e16-4222-a7d5-975db1aca215

+++ trunk/web/concrete/controllers/dashboard/groups.php
@@ -8,7 +8,7 @@ class DashboardGroupsController extends Controller {
 	
 	}	
 	
-	public function delete($delGroupId){
+	public function delete($delGroupId, $token = ''){
 
 		$u=new User();
 		try {
@@ -21,6 +21,11 @@ class DashboardGroupsController extends Controller {
 			
 			if(!($group instanceof Group)) {
 				throw new Exception(t('Invalid group ID.'));
+			}
+			
+			$valt = Loader::helper('validation/token');
+			if (!$valt->validate('delete_group_' . $delGroupId, $token)) {
+				throw new Exception($valt->getErrorMessage());
 			}
 			
 			$group->delete();
